add devdb as an option for import-tool (#951)
- adjust import scripts and crontab for devdb import tool

+++ src/import-tool.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2021 Memorial Sloan Kettering Cancer Center.
+/* Copyright (c) 2021, 2022 Memorial Sloan Kettering Cancer Center.
  *
  * This library is distributed in the hope that it will be useful, but
  * WITHOUT ANY WARRANTY, WITHOUT EVEN THE IMPLIED WARRANTY OF
@@ -97,6 +97,11 @@ static string KILL_HGNC_IMPORT_TRIGGER_FILENAME(IMPORT_TRIGGER_BASEDIR + "/hgnc-
 static string HGNC_IMPORT_IN_PROGRESS_FILENAME(IMPORT_TRIGGER_BASEDIR + "/hgnc-import-in-progress");
 static string HGNC_IMPORT_KILLING_FILENAME(IMPORT_TRIGGER_BASEDIR + "/hgnc-import-killing");
 static string HGNC_IMPORT_LOG_FILENAME(IMPORT_LOG_BASEDIR + "/hgnc-importer.log");
+static string START_DEVDB_IMPORT_TRIGGER_FILENAME(IMPORT_TRIGGER_BASEDIR + "/devdb-import-start-request");
+static string KILL_DEVDB_IMPORT_TRIGGER_FILENAME(IMPORT_TRIGGER_BASEDIR + "/devdb-import-kill-request");
+static string DEVDB_IMPORT_IN_PROGRESS_FILENAME(IMPORT_TRIGGER_BASEDIR + "/devdb-import-in-progress");
+static string DEVDB_IMPORT_KILLING_FILENAME(IMPORT_TRIGGER_BASEDIR + "/devdb-import-killing");
+static string DEVDB_IMPORT_LOG_FILENAME(IMPORT_LOG_BASEDIR + "/devdb-importer.log");
 
 /* used for command line argument checking */
 static vector<string> RECOGNIZED_IMPORTERS;
@@ -105,6 +110,7 @@ static vector<string> RECOGNIZED_COMMANDS;
 void initialize_static_objects() {
     RECOGNIZED_IMPORTERS.push_back("triage");
     RECOGNIZED_IMPORTERS.push_back("hgnc");
+    RECOGNIZED_IMPORTERS.push_back("devdb");
     RECOGNIZED_COMMANDS.push_back("start");
     RECOGNIZED_COMMANDS.push_back("kill");
     RECOGNIZED_COMMANDS.push_back("status");
@@ -129,7 +135,7 @@ vector<char> character_vector_from_string(string s) {
 void print_usage(string program_name) {
     vector<char> program_name_vector = character_vector_from_string(program_name);
     cerr << "Usage: " << basename(&program_name_vector[0]) << " importer_name command [extra_arguments]" << endl;
-    cerr << "       importer_name must be \"triage\" or \"hgnc\"" << endl;
+    cerr << "       importer_name must be \"triage\" or \"hgnc\" or \"devdb\"" << endl;
     cerr << "       valid commands:" << endl;
     cerr << "           start : requests that an import run begins as soon as possible - this may wait for an import in progress to finish before starting" << endl;
     cerr << "           kill : requests that any import in progress be halted and that any requested start be canceled" << endl;
@@ -376,6 +382,9 @@ int request_importer_start(string importer) {
     if (importer == "hgnc") {
         return request_importer_start(importer, START_HGNC_IMPORT_TRIGGER_FILENAME, KILL_HGNC_IMPORT_TRIGGER_FILENAME, HGNC_IMPORT_KILLING_FILENAME);
     }
+    if (importer == "devdb") {
+        return request_importer_start(importer, START_DEVDB_IMPORT_TRIGGER_FILENAME, KILL_DEVDB_IMPORT_TRIGGER_FILENAME, DEVDB_IMPORT_KILLING_FILENAME);
+    }
     cerr << "Error : unrecognized importer " << importer << " during attempt to start import." << endl;
     return 1;
 }
@@ -386,6 +395,9 @@ int request_importer_kill(string importer) {
     }
     if (importer == "hgnc") {
         return create_trigger_file(KILL_HGNC_IMPORT_TRIGGER_FILENAME);
+    }
+    if (importer == "devdb") {
+        return create_trigger_file(KILL_DEVDB_IMPORT_TRIGGER_FILENAME);
     }
     cerr << "Error : unrecognized importer " << importer << " during attempt to kill import." << endl;
     return 1;
@@ -432,6 +444,9 @@ int report_importer_status(string importer) {
     }
     if (importer == "hgnc") {
         return report_importer_status(importer, START_HGNC_IMPORT_TRIGGER_FILENAME, KILL_HGNC_IMPORT_TRIGGER_FILENAME, HGNC_IMPORT_IN_PROGRESS_FILENAME, HGNC_IMPORT_KILLING_FILENAME, HGNC_IMPORT_LOG_FILENAME);
+    }
+    if (importer == "devdb") {
+        return report_importer_status(importer, START_DEVDB_IMPORT_TRIGGER_FILENAME, KILL_DEVDB_IMPORT_TRIGGER_FILENAME, DEVDB_IMPORT_IN_PROGRESS_FILENAME, DEVDB_IMPORT_KILLING_FILENAME, DEVDB_IMPORT_LOG_FILENAME);
     }
     cerr << "Error : unrecognized importer " << importer << " during attempt to report status." << endl;
     return 1;
@@ -495,6 +510,9 @@ int report_importer_log(string importer, vector<string> & extra_args) {
     }
     if (importer == "hgnc") {
         return report_importer_log(importer, HGNC_IMPORT_LOG_FILENAME, extra_args);
+    }
+    if (importer == "devdb") {
+        return report_importer_log(importer, DEVDB_IMPORT_LOG_FILENAME, extra_args);
     }
     cerr << "Error : unrecognized importer " << importer << " during attempt to report log." << endl;
     return 1;
